requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.1
streamlit>=1.36.0
plotly>=5.22.0
textblob>=0.17.1

# Data stack (conditional for Python versions)
pandas>=2.2.0; python_version < "3.13"
numpy>=1.26.0; python_version < "3.13"

pandas>=2.3.0; python_version >= "3.13"
numpy>=2.1.0; python_version >= "3.13"
